Move blank node tests to the prec impl folder
