+++ test/index.js
@@ -9,4 +9,3 @@ require("./dataset/DatasetCore.test.js")({
     blankNode: N3.DataFactory.blankNode,
     dataset: t => new (require('../src/dataset/index.js'))(t)
 });
-
